+++ front-end/app/_layout.jsx
@@ -1,22 +1,21 @@
 import { Stack } from 'expo-router';
-import { AuthProvider, useAuthContext } from '../context/AuthContext';
+import { AuthProvider } from '../context/AuthContext';
+
+export const RootStack = () => {
+    return (
+        <Stack>
+            <Stack.Screen name="index" options={{ headerShown: false }} />
+            <Stack.Screen name="home" options={{ headerShown: false }} />
+            <Stack.Screen name="(auth)" options={{ headerShown: false }} />
+        </Stack>
+    );
+}
 
 const RootLayout = () => {
     return (
         <AuthProvider>
             <RootStack></RootStack>
         </AuthProvider>
-    );
-}
-
-export const RootStack = () => {
-    const { authState } = useAuthContext();
-    return (
-        <Stack>
-            {authState.authenticated ?
-                <Stack.Screen name="index" options={{ headerShown: false }} />
-                : <Stack.Screen name="(auth)" options={{ headerShown: false }} />}
-        </Stack>
     );
 }
